% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{cola_layout}
\alias{cola_layout}
\title{Title}
\usage{
cola_layout(cytoscape, name = "cola", animate = TRUE, refresh = 1,
  maxSimulationTime = 4000, ungrabifyWhileSimulating = FALSE, fit = TRUE,
  padding = 30, boundingBox = NULL, nodeDimensionsIncludeLabels = FALSE,
  ready = JS("function(){}"), stop = JS("function(){}"),
  randomize = FALSE, avoidOverlap = TRUE, handleDisconnected = TRUE,
  nodeSpacing = JS("function( node ){ return 10; }"), flow = NULL,
  alignment = NULL, gapInequalities = NULL, edgeLength = NULL,
  edgeSymDiffLength = NULL, edgeJaccardLength = NULL, unconstrIter = NULL,
  userConstIter = NULL, allConstIter = NULL, infinite = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
