% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{cytoscape}
\alias{cytoscape}
\title{<Add Title>}
\usage{
cytoscape(df = NULL, nodes = NULL, edges = NULL, id = "id",
  source = "source", target = "target", message = "testing",
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{df}{data.frame or tibble}

\item{nodes}{data.frame or tibble}

\item{edges}{data.frame or tibble}

\item{id}{character column name where df used}

\item{source}{character column name where df used}

\item{target}{character column name where df used}

\item{elementId}{}
}
\description{
<Add Description>
}
\examples{

 nodes <- data.frame(id = c('a','b'))
 edges <- data.frame(id = 'ab', source = 'a', target = 'b')

 cytoscape(nodes = nodes, edges = edges)

}
