% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcose.R
\name{fcose_layout}
\alias{fcose_layout}
\title{Cola Layout}
\usage{
fcose_layout(
  cytoscape,
  name = "fcose",
  quality = "default",
  randomize = TRUE,
  animate = TRUE,
  animationDuration = 1000,
  animationEasing = NULL,
  fit = TRUE,
  padding = 30,
  nodeDimensionsIncludeLabels = FALSE,
  uniformNodeDimensions = FALSE,
  packComponents = TRUE,
  samplingType = TRUE,
  sampleSize = 25,
  nodeSeparation = 75,
  piTol = 1e-07,
  nodeRepulsion = 4500,
  idealEdgeLength = 50,
  edgeElasticity = 0.45,
  nestingFactor = 0.1,
  numIter = 2500,
  tile = TRUE,
  tilingPaddingVertical = 10,
  tilingPaddingHorizontal = 10,
  gravity = 0.25,
  gravityRangeCompound = 1.5,
  gravityCompound = 1,
  gravityRange = 3.8,
  initialEnergyOnIncremental = 0.3,
  ready = htmlwidgets::JS("() => {}"),
  stop = htmlwidgets::JS("() => {}"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Cola Layout
}
