% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{node_style}
\alias{node_style}
\title{Node Style}
\usage{
node_style(cytoscape, ...)
}
\arguments{
\item{cytoscape}{object}

\item{...}{additional parameters passed to the style node object}
}
\value{
ctytoscape object
}
\description{
Any recognised cytoscape style option can be passed to the node style. For available
options see the \href{http://js.cytoscape.org/#style/node-body}{cytoscape documentation}.
}
\examples{

nodes <- data.frame(id = c('a','b'))
edges <- data.frame(id = 'ab', source = 'a', target = 'b')

cytoscape(nodes = nodes, edges = edges) \%>\% node_style('background-color' = '#ff0000')

}
