% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panzoom.R
\name{panzoom}
\alias{panzoom}
\title{panzoom Layout}
\usage{
panzoom(cytoscape, name = "panzoom", zoomFactor = 0.05, zoomDelay = 45,
  minZoom = 0.1, maxZoom = 10, fitPadding = 50, panSpeed = 10,
  panDistance = 10, panDragAreaSize = 75, panMinPercentSpeed = 0.25,
  panInactiveArea = 8, panIndicatorMinOpacity = 0.5, zoomOnly = FALSE,
  fitSelector = NULL,
  animateOnFit = htmlwidgets::JS("function(){return false;}"),
  fitAnimationDuration = 1000, sliderHandleIcon = "fa fa-minus",
  zoomInIcon = "fa fa-plus", zoomOutIcon = "fa fa-minus",
  resetIcon = "fa fa-expand", ...)
}
\arguments{
\item{zoomDelay}{}

\item{...}{}
}
