% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{cytoscape}
\alias{cytoscape}
\title{Cytoscape Network Chart}
\usage{
cytoscape(df = NULL, nodes = NULL, edges = NULL, json = NULL,
  id = "id", source = "source", target = "target", message = NULL,
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{df}{data.frame or tibble}

\item{nodes}{data.frame or tibble}

\item{edges}{data.frame or tibble}

\item{json}{pass single json object. This option over-rules all others.}

\item{id}{character column name where df used}

\item{source}{character column name where df used}

\item{target}{character column name where df used}

\item{elementId}{}
}
\description{
Cytoscape Network Chart
}
\examples{

 nodes <- data.frame(id = c('a','b'))
 edges <- data.frame(id = 'ab', source = 'a', target = 'b')

 cytoscape(nodes = nodes, edges = edges)

}
