% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{edge_style}
\alias{edge_style}
\title{Edge Style}
\usage{
edge_style(cytoscape, ...)
}
\arguments{
\item{cytoscape}{object}

\item{...}{additional parameters passed to the style edge object}
}
\description{
Any recognised cytoscape style option can be passed to the node style. For available
options see the \href{http://js.cytoscape.org/#style/edge-line}{cytoscape documentation}.
}
\examples{

nodes <- data.frame(id = c('a','b'))
edges <- data.frame(id = 'ab', source = 'a', target = 'b')

cytoscape(nodes = nodes, edges = edges) \%>\% edge_style('line-color' = '#ff0000')

}
